+++ solc/CommandLineInterface.cpp
@@ -157,6 +157,8 @@ static string const g_strOutputDir = "output-dir";
 static string const g_strOverwrite = "overwrite";
 static string const g_strRevertStrings = "revert-strings";
 static string const g_strStorageLayout = "storage-layout";
+static string const g_strStopAfter = "stop-after";
+static string const g_strParsing = "parsing";
 
 /// Possible arguments to for --revert-strings
 static set<string> const g_revertStringsArgs
@@ -314,6 +316,17 @@ static bool needsHumanTargetedStdout(po::variables_map const& _args)
 		if (_args.count(arg))
 			return true;
 	return false;
+}
+
+bool CommandLineInterface::checkMutuallyExclusive(std::string const& _optionA, std::string const& _optionB) const
+{
+	if (m_args.count(_optionA) && m_args.count(_optionB))
+	{
+		serr() << "Option " << _optionA << " and " << _optionB << " are mutually exclusive." << endl;
+		return false;
+	}
+
+	return true;
 }
 
 void CommandLineInterface::handleBinary(string const& _contract)
@@ -800,6 +813,11 @@ General Information)").c_str(),
 			po::value<string>()->value_name(boost::join(g_revertStringsArgs, ",")),
 			"Strip revert (and require) reason strings or add additional debugging information."
 		)
+		(
+			g_strStopAfter.c_str(),
+			po::value<string>()->value_name("stage"),
+			"Stop execution after the given compiler stage. Valid options: \"parsing\"."
+		)
 	;
 	desc.add(outputOptions);
 
@@ -992,11 +1010,23 @@ General Information)").c_str(),
 		return false;
 	}
 
-	if (m_args.count(g_argColor) && m_args.count(g_argNoColor))
-	{
-		serr() << "Option " << g_argColor << " and " << g_argNoColor << " are mutualy exclusive." << endl;
+	if (!checkMutuallyExclusive(g_argColor, g_argNoColor))
 		return false;
-	}
+
+	static vector<string> const conflictingWithStopAfter{
+		g_argBinary,
+		g_argIR,
+		g_argIROptimized,
+		g_argEwasm,
+		g_argGas,
+		g_argAsm,
+		g_argAsmJson,
+		g_argOpcodes
+	};
+
+	for (auto& option: conflictingWithStopAfter)
+		if (!checkMutuallyExclusive(g_strStopAfter, option))
+			return false;
 
 	m_coloredOutput = !m_args.count(g_argNoColor) && (isatty(STDERR_FILENO) || m_args.count(g_argColor));
 
@@ -1132,6 +1162,17 @@ bool CommandLineInterface::processInput()
 				filesystem_path.remove_filename();
 			m_allowedDirectories.push_back(filesystem_path);
 		}
+	}
+
+	if (m_args.count(g_strStopAfter))
+	{
+		if (m_args[g_strStopAfter].as<string>() != "parsing")
+		{
+			serr() << "Valid options for --" << g_strStopAfter << " are: \"parsing\".\n";
+			return false;
+		}
+		else
+			m_stopAfter = CompilerStack::State::Parsed;
 	}
 
 	vector<string> const exclusiveModes = {
@@ -1413,7 +1454,7 @@ bool CommandLineInterface::processInput()
 				m_compiler->setParserErrorRecovery(true);
 		}
 
-		bool successful = m_compiler->compile();
+		bool successful = m_compiler->compile(m_stopAfter);
 
 		for (auto const& error: m_compiler->errors())
 		{
@@ -1863,7 +1904,10 @@ void CommandLineInterface::outputCompilationResults()
 	handleAst(g_argAstJson);
 	handleAst(g_argAstCompactJson);
 
-	if (!m_compiler->compilationSuccessful())
+	if (
+		!m_compiler->compilationSuccessful() &&
+		m_stopAfter == CompilerStack::State::CompilationSuccessful
+	)
 	{
 		serr() << endl << "Compilation halted after AST generation due to errors." << endl;
 		return;