+++ test/cmdlineTests.sh
@@ -437,6 +437,9 @@ SOLTMPDIR=$(mktemp -d)
 )
 rm -rf "$SOLTMPDIR"
 
+printTask "Testing AST export with stop-after=parse..."
+"$REPO_ROOT/test/stopAfterParseTests.sh"
+
 printTask "Testing soljson via the fuzzer..."
 SOLTMPDIR=$(mktemp -d)
 (